pyproject.toml
[project]
name = "chatbot-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "openai-agents>=0.0.15",
]
